Enhance checking in some operator. test=develop

+++ python/paddle/fluid/tests/unittests/test_layers.py
@@ -3000,8 +3000,7 @@ class TestBook(LayerTest):
             out, ids = layers.argsort(input=data, axis=1)
 
             theta = layers.data(name="theta", shape=[2, 3], dtype="float32")
-            out_shape = layers.data(
-                name="out_shape", shape=[-1], dtype="float32")
+            out_shape = layers.data(name="out_shape", shape=[-1], dtype="int32")
             data_0 = layers.affine_grid(theta, out_shape)
             data_1 = layers.affine_grid(theta, [5, 3, 28, 28])
